+++ webknossos/tests/conftest.py
@@ -1,5 +1,4 @@
 from os import makedirs
-from pathlib import Path
 from shutil import rmtree
 from typing import Any, Dict, Generator, Union
 from unittest.mock import MagicMock, patch
@@ -11,7 +10,7 @@ from vcr.stubs import httpx_stubs
 
 from webknossos.client.context import _clear_all_context_caches
 
-TESTOUTPUT_DIR = Path("testoutput")
+from .constants import TESTOUTPUT_DIR
 
 
 @pytest.fixture(autouse=True, scope="function")